fix out of order insert for states related to a process

+++ vino/src/main/java/it/unimol/vino/models/request/AddStateToProcessRequest.java
@@ -17,7 +17,7 @@ public class AddStateToProcessRequest {
     private Long processId;
 
     @NonNull
-    private Long sequence;
+    private Integer sequence;
 
     private Long waste = null;
